requirements.txt
viam-sdk
openai
pyyaml